refactor(core): standardize function signatures and fix bugs in execution and expansion
- Update `UTILS_SRC` and `STDFCTS_SRC` to incorporate new source files and reformat the lists for improved project maintainability.
- Correct the argument order in the `join_two_strings` call within `find_in_path` to ensure proper command path construction.
- Refactor `expand_double_quotes_word` to simplify the main expansion loop and remove an unused variable for cleaner code.
- Streamline the loop termination condition in `expand_single_quote_word` by removing a redundant null-terminator check.
- Modify function prototypes by removing `const` qualifiers from string-handling functions and adding the declaration for `ft_min`.
- Adjust the `skip_spaces` function prototype to return `size_t` instead of `int` for better type consistency with string indices.
- Remove the `const` qualifier from the source pointer parameter in the `ft_memcpy` implementation to align with its new prototype.
- Update `ft_split` and its internal helper functions by removing the `const` qualifier from string parameters to match the new header definition.
- Introduce `smalloc.c` to provide a safe `malloc` wrapper that includes built-in error handling and exits upon allocation failure.

+++ includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: yjaafar <yjaafar@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 08:41:01 by yjaafar           #+#    #+#             */
-/*   Updated: 2025/08/07 05:03:32 by yjaafar          ###   ########.fr       */
+/*   Updated: 2025/08/08 00:57:03 by yjaafar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ void            _free(char **str);
 t_list_info	    *init_list_info_struct();
 enum e_operator	check_token(char *token);
 t_list			*creat_node(void *content);
-int				skip_spaces(char *str, size_t i);
+size_t			skip_spaces(char *str, size_t i);
 bool			check_if_match(char *sep, char c);
 void			join_list(t_list_info *lst_1, t_list_info *lst_2);
 void			list_add_back(t_list_info *list_info, t_list *node);
